Setup and draw the snake for the game
In the Gameplay class implement actionlistener and keylistener and implement their methods.  Create arrays to hold the snakes x and y coordinates, and create Timer object. After all of that is done create boolean checks for 4 directions it can move in and in paint method draw the snake in the intial move state.