Create right movement right snake
In Gameplay class implement the checks to see which direction we are in the keypressed method. In the actionperformed method check see if moving in right is true and if so set it new coordinates based on its previous state and body state and move. Afterwards draw the snake in its new position.