Add enemy and score
Create arraylist to hold the point x position and yposition and fill up the arrays with all the possible locations it can spawn. Create a way for the snake to grow when it collides with the point and draw a score counter and a length counter.
Note: Small bug index will be out of bounds if snake gets too long, error could be with how i handled the coordinates.